api/requirements.txt
fastapi
uvicorn[standard]
pydantic
langchain
langchain-openai
langchain-google-genai
langchain-community
langchain-perplexity
langchain-text-splitters
langchain-pinecone
openai
google-generativeai
faiss-cpu
pinecone[asyncio]
python-dotenv
requests
lxml
tavily-python
html2text
beautifulsoup4
azure-messaging-webpubsubservice
aiohttp